% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{Alien.Vault.Reputation}
\alias{Alien.Vault.Reputation}
\title{Retrieves Alien Vault's IP reputation database}
\usage{
Alien.Vault.Reputation(refresh = FALSE,
  alien.vault.reputation.url = "http://reputation.alienvault.com/reputation.data")
}
\arguments{
\item{refresh}{refresh the database? (bool)}

\item{alien.vault.reputation.url}{URL of the AlienVault data (chr) -
defaults to \code{http://reputation.alienvault.com/reputation.data}}
}
\value{
data.table with IP & Reputation information.
  \itemize{
    \item \code{IP} - IPv4 address
    \item \code{Risk} - how risky is the target (1-10)
    \item \code{Reliability} - how reliable is the rating (1-10)
    \item \code{Activity} - what type of host is it
    \item \code{Country} - what is the IPv4 country of origin
    \item \code{City} - what is the IPv4 city of origin
    \item \code{Latitude} - geolocated latitude of the IPv4
    \item \code{Longitude} - geolocated longitude of the IPv4
  }
}
\description{
Retrieves Alien Vault's IP reputation database.
}
\details{
AlienValut refreshes every hour, but the onus is on the caller to force a
refresh. First-time call will setup a cache directory & file in the user's
home directory, download & generate the data frame then write the data frame
out as an R object. Future calls will just re-read this data frame unless
\code{refresh == TRUE} should the function refresh the database.

Please be kind to the AlienValut servers & only refresh if you really need to.
}
\seealso{
\itemize{
    \item Background on AlienValut's IP rep db: \url{http://labs.alienvault.com/labs/index.php/projects/open-source-ip-reputation-portal/download-ip-reputation-database/}
    \item More info on AlienVault's database: \url{http://www.slideshare.net/alienvault/building-an-ip-reputation-engine-tracking-the-miscreants}
  }
}

