% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BulkOrigin}
\alias{BulkOrigin}
\title{Retrieves BGP Origin ASN info for a list of IPv4 addresses}
\usage{
BulkOrigin(ip.list, host = "v4.whois.cymru.com", port = 43)
}
\arguments{
\item{ip.list}{vector of IPv4 address (character - dotted-decimal)}

\item{host}{which server to perform the lookup (chr) -
defaults to \code{v4.whois.cymru.com}}

\item{port}{TCP port to use to connect to \code{host} (int) -
defaults to port \code{43}}
}
\value{
data frame of BGP Origin ASN lookup results
  \itemize{
    \item \code{AS} - AS #
    \item \code{IP} - IPv4 (passed in)
    \item \code{BGP.Prefix} - BGP CIDR
    \item \code{CC} - Country code
    \item \code{Registry} - Registry it falls under
    \item \code{Allocated} - date it was allocated
    \item \code{AS.Name} - AS name
  }
}
\description{
Returns a list (slots are named by the input IPv4 addresses)
             with lookup results per slot
}
\note{
The Team Cymru's service is NOT a GeoIP service! Do not use this
      function for that as your results will not be accurate.
}
\seealso{
\url{http://www.team-cymru.org/Services/}
}

