% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{BulkPeer}
\alias{BulkPeer}
\title{Retrieves BGP Peer ASN info for a list of IPv4 addresses}
\usage{
BulkPeer(ip.list, host = "v4-peer.whois.cymru.com", port = 43)
}
\arguments{
\item{ip.list}{vector of IPv4 address (character - dotted-decimal)}

\item{host}{which server to perform the lookup (chr) -
defaults to \code{v4.whois.cymru.com}}

\item{port}{TCP port to use to connect to \code{host} (int) -
defaults to \code{43}}
}
\value{
data frame of BGP Peer ASN lookup results
  \itemize{
    \item \code{Peer.AS} - peer AS #
    \item \code{IP} - IPv4 (passsed in)
    \item \code{BGP.Prefix} - BGP CIDR block
    \item \code{CC} - Country code
    \item \code{Registry} - Registry it falls under
    \item \code{Allocated} - date allocated
    \item \code{Peer.AS.Name} - peer name
  }
}
\description{
Retrieves BGP Peer ASN info for a list of IPv4 addresses
}
\note{
The Team Cymru's service is NOT a GeoIP service! Do not use this
      function for that as your results will not be accurate.
}
\seealso{
\url{http://www.team-cymru.org/Services/}
}

