% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{package}
\name{netintel-package}
\alias{netintel}
\alias{netintel-package}
\title{netintel - a package containing various function to support IPv4 address, ASN and general node intelligence lookups}
\description{
When attempting to uncover "badness", a security data scientist,
intelligence/ops analyst or incident responder needs metadata on the IP
addresses, hosts, domains, URLs, etc the come across. This package provides
functions that aid in collecting this metadata from within R.
}
\author{
Bob Rudis <\url{http://github.com/hrbrmstr} and David Severski <\url{https://github.com/davidski}>
}
\seealso{
The latest/development version of the package is on github - \url{http://github.com/hrbrmstr/netintel} where there are examples of usage.
}

