% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{Nothink.Blocklist}
\alias{Nothink.Blocklist}
\title{Retrieves Nothink Malware DNS network traffic blacklist (IP/FQDN)}
\usage{
Nothink.Blocklist(refresh = FALSE,
  nothink_url = "http://www.nothink.org/blacklist/blacklist_malware_dns.txt")
}
\arguments{
\item{refresh}{refresh the database? (bool)}

\item{nothink_url}{Nothink blacklist URL (chr) -
defaults to \code{http://www.nothink.org/blacklist/blacklist_malware_dns.txt}}
}
\value{
List of two singe-column data frames, one for each blocklist
  \itemize{
    \item \code{domains} - Zeus domains (column name: \code{domain})
    \item \code{ips} - Zeus ips (column name: \code{IP})
}
}
\description{
Retrieves Nothink Malware DNS network traffic blacklist (IP/FQDN)
}
\details{
The Nothink blocklist refreshes regularly, but the onus is on the caller to force a
refresh. First-time call will setup a cache directory & file in the user's
home directory, download & generate the data frame then write the data frame
out as an R object. Future calls will just re-read this data frame unless
\code{refresh == TRUE} should the function refresh the database.
}
\seealso{
Nothink - \url{http://www.nothink.org/}
}

