% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{Zeus.Blocklist}
\alias{Zeus.Blocklist}
\title{Retrieves Zeus Blocklist (IP/FQDN/URL)}
\usage{
Zeus.Blocklist(refresh = FALSE,
  domains_url = "https://zeustracker.abuse.ch/blocklist.php?download=domainblocklist",
  ips_url = "https://zeustracker.abuse.ch/blocklist.php?download=ipblocklist",
  urls_url = "https://zeustracker.abuse.ch/blocklist.php?download=compromised")
}
\arguments{
\item{refresh}{refresh the database? (bool)}

\item{domains_url}{Zeus domains blocklist URL (chr) -
defaults to \code{https://zeustracker.abuse.ch/blocklist.php?download=domainblocklist}}

\item{ips_url}{Zeus IP blocklist URL (chr) -
defaults to \code{https://zeustracker.abuse.ch/blocklist.php?download=ipblocklist}}

\item{urls_url}{Zeus compromised URLs blocklist URL (chr) -
defaults to \code{https://zeustracker.abuse.ch/blocklist.php?download=compromised}}
}
\value{
List of three singe-column data frames, one for each blocklist
  \itemize{
    \item \code{domains} - Zeus domains (column name: \code{domain})
    \item \code{ips} - Zeus ips (column name: \code{IP})
    \item \code{urls} - Zeus domains (column name: \code{URL})
}
}
\description{
Retrieves Zeus Blocklist (IP/FQDN/URL)
}
\details{
The Zeus blocklist refreshes regularly, but the onus is on the caller to force a
refresh. First-time call will setup a cache directory & file in the user's
home directory, download & generate the data frame then write the data frame
out as an R object. Future calls will just re-read this data frame unless
\code{refresh == TRUE} should the function refresh the database.
}
\seealso{
Zeus blocklist info - \url{https://zeustracker.abuse.ch/blocklist.php}
}

