% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{CIRCL.BGP.Rank}
\alias{CIRCL.BGP.Rank}
\title{Retrieves CIRCL aggregated, historical/current BGP rank data}
\usage{
CIRCL.BGP.Rank(asn.list, circl.base.url = "http://bgpranking.circl.lu/csv/")
}
\arguments{
\item{asn.list}{character vector of ASN ids (character)}

\item{circl.base.url}{CIRCL server base URL (chr) -
defaults to \code{http://bgpranking.circl.lu/csv/}}
}
\value{
data frame of CIRCL rank data
  \itemize{
    \item \code{asn} asn #
    \item \code{day} date
    \item \code{rank} current rank that day
  }
}
\description{
Retrieves CIRCL aggregated, historical/current BGP rank data
}
\examples{
CIRCL.BGP.Rank(57954)
}
\seealso{
\itemize{
    \item Background on CIRCL Project (+source) \url{https://github.com/CIRCL/bgp-ranking}
    \item CIRCL BGP site \url{http://bgpranking.circl.lu/}
  }
}

