% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SANS.ASN.Detail}
\alias{SANS.ASN.Detail}
\title{Retrieves SANS ASN intel currently tracked IP detail}
\usage{
SANS.ASN.Detail(asn,
  sans.base.url = "http://isc.sans.edu/asdetailsascii.html?as=")
}
\arguments{
\item{asn}{ASN to lookup (character) - no \code{AS} prefix}

\item{sans.base.url}{SANS server base URL (chr) - defaults to
\code{http://isc.sans.edu/asdetailsascii.html?as=}}
}
\value{
data frame of SANS ASN IP data
  \itemize{
    \item \code{Source.IP} is 0 padded so each byte is three digits long
    \item \code{Reports.Count} number of packets received
    \item \code{Targets.Count} number of target IPs that reported packets from this source
    \item \code{First.Seen} First time we saw a packet from this source
    \item \code{Last.Seen} Last time we saw a packet from this source
    \item \code{Updated.Date.Time} Last date+time the record was updated
  }
}
\description{
Retrieves SANS ASN intel currently tracked IP detail
}
\note{
IPs are removed if they have not been seen in 30 days.
}
\seealso{
\url{https://isc.sans.edu/as.html}
}

